+++ packages/i18n/src/test/index.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { dcnpgettext, sprintf, __, _x, _n, _nx, setLocaleData } from '../';
+import { sprintf, __, _x, _n, _nx, setLocaleData } from '../';
 
 // Mock memoization as identity function. Inline since Jest errors on out-of-
 // scope references in a mock callback.
@@ -34,13 +34,13 @@ const additionalLocaleData = {
 setLocaleData( localeData, 'test_domain' );
 
 describe( 'i18n', () => {
-	describe( 'dcnpgettext()', () => {
-		it( 'absorbs errors', () => {
-			const result = dcnpgettext( 'domain-without-data', undefined, 'Hello' );
+	it( 'absorbs errors', () => {
+		__( 'Hello', 'domain-without-data' );
+		_x( 'feed', 'verb', 'domain-without-data' );
+		_n( '%d banana', '%d bananas', 3, 'domain-without-data' );
+		_nx( '%d apple', '%d apples', 3, 'fruit', 'domain-without-data' );
 
-			expect( console ).toHaveErrored();
-			expect( result ).toBe( 'Hello' );
-		} );
+		expect( console ).toHaveErrored();
 	} );
 
 	describe( '__', () => {